first round of paths working (#27)
